A3 Cloud VM Automation/requirements.txt
python==3.9.7
azure-cli==2.30.0
Google Cloud SDK == 363.0.0